Created Product Model
